pyproject.toml
[tool.poetry]
name = "dev-bootstrap"
version = "0.1.0"
description = "Setup your Git Machine"
authors = ["Edgar Treischl <Edgar.Treischk@isb.bayern.de>"]
readme = "README.md"
packages = [
    { include = "devbootstrap", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.14,<3.15"
pyyaml = "^6.0"
typer = "^0.24.1"
pandas = "^3.0.1"
python-dotenv = "^1.2.2"

[tool.poetry.scripts]
dev = "devbootstrap.cli:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "mkdocs-material (>=9.7.5,<10.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
